debug tensor_make_padded_square and make unit tests
There were a couple problems with this function related
to indexing on the edges.  I'm not exactly sure what
particular thing was wrong, but I split the cases up
and cached a target_idx and src_idx because it made
things easier to understand.  Although the code is
more verbose, and someone would be welcome to remerge
all the else if statements into a single one again,
this makes it far easier to debug.

The unit tests may seem redundant, but I added them
anyway because it seemed like at least initially, the
larger the input I kept sending in, the more I detected
issues with the indexing.  I believe I have fixed this
now.

Signed-off-by: Christopher Goebel <cmgoebel@iupui.edu>