[wip] naive version and bugs in second version (commented)
the fixes in this primarily focused on the cpu code in
layer_conv.c

Part of the fixes is commented out at the bottom of the
file, but that portion of the code is related to the
next (slightly less naive) version.  In this commit,
I focus on the "most naive version" of col2im inner.

The operation may still not be correct, as there
is a potential race condition that I have not
addressed in the inner loop of the GPU code.  Right
now, I have only executed the GPU code with a single
thread.  This proves that the translation from the
cpu code to the GPU code is correct, but not that the
GPU code will remain correct under parallel execution.

Signed-off-by: Christopher Goebel <cmgoebel@iu.edu>