adds conv forward device harness (working) and related
Currently the convolution_device_forward... group of
functions is a bit loose as to what their responsibility
should be.  This commit represents the combination of
all the individual device functions related to the
forward into a single function (or several).

A new test was added which mimics the convolution
forward test on the host, but calls the device
harness.

Currently, memory should be host size, and the output
of the layer is returned as host mem, while the cache
is kept in the device.

Signed-off-by: Christopher Goebel <cmgoebel@iu.edu>