Backward gpu (#46)
* adds all basics (untested) of convolution_backward_device

This is a significant commit with many changes in the
dev_layer_conv.cu file which relate to the
convolution_backward_device.  Although the tests were
not run, and are likely breaking.  This commit includes
a (at least partially) running function.

Signed-off-by: Christopher Goebel <cmgoebel@iu.edu>

* fixes bugs in backward, adds some basic functions

There were a lot of bugs in backward-device stuff.  This
commit fixes all of the ones that were causing the tests
to fail at least.

Signed-off-by: Christopher Goebel <cmgoebel@iu.edu>