touch up device forward host harness and forward
This commit changes the way the host harness works in
the convolution forward process.  It additionally
changes the forward itself to reduce the memory
usage, and to make it a pure device memory function.

The point of this commit is to improve the forward
in preparation for a looping forward backward test
for speed.

Modify tests to reflect difference.  Additionally
adds the DEBUG_GPU_FUNCTION_ENTRY_NOTIFICATIONS
macro.  Doesn't seem to be working, but added it
anyway.

Signed-off-by: Christopher Goebel <cmgoebel@iu.edu>