[enhancement]: adds harness for device utils
moves some things around, but creates harness functions
as an example of what all our call structures will need
to be in the future in order to gain control over the
block and thread sizes.

Note that this should be further extended to actually
accept dim3, but for now it will not be.

Signed-off-by: Christopher Goebel <cmgoebel@iu.edu>