fix all tensor device alloc dealloc counters and prints
Changes the device allocation and deallocation to take in
real tensors which allow me to do a better job forcing
everything to notify on its allocation and deallocation
counter.

Should help figure out whats up with the memory leaks in
convolution forward and backward.

Also further changes toward unifying on in in the whole
project.

Signed-off-by: Christopher Goebel <cmgoebel@iu.edu>